Removing duplicated Package.swift
